+++ components/src/lib/index.ts
@@ -1,2 +1,5 @@
-export { Modal, type Props } from './components/common/Modal';
+export { Modal, type ModalProps } from './components/common';
 export { useModal } from './hooks/useModal';
+export { default as AlertModal } from './components/AlertModal';
+export { default as ConfirmModal } from './components/ConfirmModal';
+export { default as PromptModal } from './components/PromptModal';
